PI-2365: Added compare endpoint
